Add Vitest test suite with MSW
