+++ package.json
@@ -5,12 +5,18 @@
     "build": "esbuild src/index.ts --bundle --outdir=public/dist --format=esm",
     "dev": "npm-run-all --parallel build:watch serve",
     "build:watch": "esbuild src/index.ts --bundle --outdir=public/dist --format=esm --watch",
-    "serve": "lite-server"
+    "serve": "lite-server",
+    "test": "vitest run"
   },
   "devDependencies": {
+    "@testing-library/dom": "^10.4.0",
+    "@testing-library/jest-dom": "^6.6.3",
+    "esbuild": "^0.25.5",
+    "jsdom": "^26.1.0",
     "lite-server": "^2.6.1",
+    "msw": "^2.10.3",
     "npm-run-all": "^4.1.5",
     "typescript": "^5.2.2",
-    "esbuild": "^0.25.5"
+    "vitest": "^3.2.4"
   }
 }
